+++ crates/frontend/src/circuits/mod.rs
@@ -1,2 +1,3 @@
 pub mod base64;
+pub mod basic;
 pub mod sha256;
